Changed whole system of moderating and changing of Status of the Product

+++ src/main/java/uz/pdp/backend/olxapp/service/ProductServiceImpl.java
@@ -3,7 +3,9 @@ package uz.pdp.backend.olxapp.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
@@ -12,17 +14,18 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import uz.pdp.backend.olxapp.entity.*;
+import uz.pdp.backend.olxapp.entity.abstractEntity.AbstractEntity;
 import uz.pdp.backend.olxapp.entity.abstractEntity.LongIdAbstract;
 import uz.pdp.backend.olxapp.enums.Role;
+import uz.pdp.backend.olxapp.enums.Status;
 import uz.pdp.backend.olxapp.exception.EntityNotFoundException;
 import uz.pdp.backend.olxapp.exception.IllegalActionException;
-import uz.pdp.backend.olxapp.mapper.AttachmentMapper;
 import uz.pdp.backend.olxapp.mapper.ProductMapper;
 import uz.pdp.backend.olxapp.payload.*;
 import uz.pdp.backend.olxapp.repository.AttachmentRepository;
 import uz.pdp.backend.olxapp.repository.CategoryRepository;
-import uz.pdp.backend.olxapp.repository.ProductImageRepository;
 import uz.pdp.backend.olxapp.repository.ProductRepository;
+import uz.pdp.backend.olxapp.specifation.ProductSpecification;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -34,17 +37,16 @@ public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
     private final ProductMapper productMapper;
     private final AttachmentService attachmentService;
-    private final AttachmentMapper attachmentMapper;
     private final CategoryRepository categoryRepository;
     private final AttachmentRepository attachmentRepository;
-    private final ProductImageRepository productImageRepository;
 
     @Override
     public PageDTO<ProductDTO> read(Integer page, Integer size) {
         Sort sort = Sort.by(LongIdAbstract.Fields.id);
         PageRequest pageRequest = PageRequest.of(page, size, sort);
 
-        Page<Product> productPage = productRepository.findAll(pageRequest);
+        //Status ACTIVE bo'lganlarni olib kelamiz
+        Page<Product> productPage = productRepository.findAllByStatus(pageRequest, Status.ACTIVE);
 
         return new PageDTO<>(
                 productPage.getContent().stream().map(productMapper::toDto).toList(),
@@ -62,17 +64,18 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public ProductDTO read(Long id) {
 
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product not found with id: " + id, HttpStatus.NOT_FOUND));
+        //Bitta productni olib kelamiz uni ham statusi ACTIVE bo'lishi kerak
+        Product product = productRepository.findByIdAndStatus(id, List.of(Status.ACTIVE))
+                .orElseThrow(() -> new EntityNotFoundException("Your product may be is not active or not found", HttpStatus.NOT_FOUND));
 
         return productMapper.toDto(product);
-
     }
 
     @Override
     public ProductDTO increaseViewCount(Long id) {
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product not found with id: " + id, HttpStatus.NOT_FOUND));
+        Product product = productRepository
+                .findByIdAndStatus(id, List.of(Status.ACTIVE))
+                .orElseThrow(() -> new EntityNotFoundException("Your product may be is not active or not found", HttpStatus.NOT_FOUND));
 
         product.setViewCounter(product.getViewCounter() + 1);
         return productMapper.toDto(productRepository.save(product));
@@ -85,9 +88,9 @@ public class ProductServiceImpl implements ProductService {
     public ProductDTO save(ProductReqDTO productReqDTO) {
 
 
-        Category category = categoryRepository.findById(productReqDTO.getCategoryId())
-                .orElseThrow(() -> new EntityNotFoundException("Category not found with id: " + productReqDTO.getCategoryId(), HttpStatus.NOT_FOUND));
+        Category category = categoryRepository.findByIdOrThrow(productReqDTO.getCategoryId());
 
+        //noinspection ExtractMethodRecommender
         if (!category.getChildren().isEmpty()) {
             throw new IllegalArgumentException("Category has child categories");
         }
@@ -103,7 +106,8 @@ public class ProductServiceImpl implements ProductService {
         product.setDescription(productReqDTO.getDescription());
         product.setPrice(productReqDTO.getPrice());
         product.setCategory(category); // yoki categoryService.getById(id)
-        product.setActive(productReqDTO.isActive());
+///        product.setActive(productReqDTO.isActive()); buni orniga endi biz Enum bilan ishlayapmiz
+        product.setStatus(Status.PENDING_REVIEW);
         product.setCreatedBy(user);
 
         Product savedProduct = productRepository.save(product);// asosiy product saqlanadi
@@ -136,8 +140,7 @@ public class ProductServiceImpl implements ProductService {
 
                 // Attachment saqlaymiz
                 AttachmentDTO attachmentDTO = attachmentService.upload(file); // bu DTO ichida ID bo'lishi kerak
-                Attachment attachment = attachmentRepository.findById(attachmentDTO.getId())
-                        .orElseThrow(() -> new RuntimeException("Attachment not found"));
+                Attachment attachment = attachmentRepository.findByIdOrElseTrow(attachmentDTO.getId());
 
                 // ProductImage ni yaratamiz
                 ProductImage image = new ProductImage();
@@ -151,22 +154,21 @@ public class ProductServiceImpl implements ProductService {
 
 
         // 3. ProductDTO qaytarish
-        return productMapper.toDto(product); // yoki qo‘lda to‘ldiring
+        return productMapper.toDto(product);
     }
 
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public ProductDTO updateProduct(Long productId, ProductUpdateDTO productUpdateDTO) {
-        // 1. Mahsulotni bazadan topamiz
-        Product product = productRepository.findById(productId)
-                .orElseThrow(() -> new EntityNotFoundException("Product not found with id: " + productId, HttpStatus.NOT_FOUND));
 
-        // 2. NULL'dan himoyalangan ro'yxatlarni olish
+        Product product = productRepository
+                .findByIdAndStatus(productId, List.of(Status.ACTIVE, Status.REJECTED))
+                .orElseThrow(() -> new EntityNotFoundException("Your product may be is not active or not found", HttpStatus.NOT_FOUND));
+
         List<ExistedImageDTO> existedImages = Optional.ofNullable(productUpdateDTO.getExistedImages()).orElse(Collections.emptyList());
         List<ProductNewImageDTO> newImages = Optional.ofNullable(productUpdateDTO.getProductNewImages()).orElse(Collections.emptyList());
 
-        // 3. SIZNING YANGI VALIDATSIYANGIZ (to'g'rilangan)
         if (existedImages.size() + newImages.size() > 8) {
             throw new IllegalActionException("Maximum 8 images are allowed.", HttpStatus.BAD_REQUEST);
         }
@@ -180,22 +182,20 @@ public class ProductServiceImpl implements ProductService {
                 .count();
 
         if (isMainCountInExistingImages + isMainCountInNewImages != 1) {
-            throw new IllegalActionException("One main image should be selected", HttpStatus.BAD_REQUEST); // Agar asosiy rasm soni 1 dan katta bo'lsa, hatolikni yuz beradigan yer
+            throw new IllegalActionException("One main image should be selected", HttpStatus.BAD_REQUEST);
         }
 
         product.getProductImages().forEach(image -> image.setMain(Boolean.FALSE));
 
         if (!newImages.isEmpty()) {
-            // Yangi rasmni yuklash uchun kodlar
             for (ProductNewImageDTO productNewImage : productUpdateDTO.getProductNewImages()) {
 
                 MultipartFile file = productNewImage.getFile();
                 AttachmentDTO upload = attachmentService.upload(file);
-                Attachment attachment = attachmentRepository.findById(upload.getId())
-                        .orElseThrow(() -> new EntityNotFoundException("Attachment not found", HttpStatus.NOT_FOUND));
+                Attachment attachment = attachmentRepository.findByIdOrElseTrow(upload.getId());
 
                 if (Boolean.TRUE.equals(productNewImage.getMain())) {
-                    product.getProductImages().forEach(image -> image.setMain(Boolean.FALSE)); // Barchasi asosiy emas
+                    product.getProductImages().forEach(image -> image.setMain(Boolean.FALSE));
                     product.getProductImages().add(new ProductImage(product, attachment, Boolean.TRUE));
                 } else {
                     product.getProductImages().add(new ProductImage(product, attachment, Boolean.FALSE));
@@ -222,36 +222,33 @@ public class ProductServiceImpl implements ProductService {
             }
         });
 
-        // 2. Xavfsizlik tekshiruvi
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (!(authentication.getPrincipal() instanceof User user)) {
             throw new AccessDeniedException("User is not authenticated");
         }
-        if (!product.getCreatedBy().getId().equals(user.getId())) {
+        if (!product.getCreatedBy().getId().equals(user.getId()) || user.getRole().equals(Role.ADMIN)) {
             throw new AccessDeniedException("You are not the owner of this product");
         }
 
         product.setTitle(productUpdateDTO.getTitle());
         product.setDescription(productUpdateDTO.getDescription());
         product.setPrice(productUpdateDTO.getPrice());
-        product.setActive(productUpdateDTO.isActive());
-        product.setIsApproved(false);
+        product.setStatus(Status.PENDING_REVIEW);
 
 
-        // 5. O'zgarishlarni saqlaymiz
         Product updatedProduct = productRepository.save(product);
 
-        // 6. Natijani DTO ko'rinishida qaytaramiz
         return productMapper.toDto(updatedProduct);
     }
 
 
     @Override
     @Transactional
-    public void updateStatus(Long id, boolean active) {
+    public void updateStatus(Long id) {
 
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product not found with id: " + id, HttpStatus.NOT_FOUND));
+        Product product = productRepository
+                .findByIdAndStatus(id, List.of(Status.ACTIVE))
+                .orElseThrow(() -> new EntityNotFoundException("Product not found", HttpStatus.NOT_FOUND));
 
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication.getPrincipal() instanceof User user) {
@@ -262,11 +259,8 @@ public class ProductServiceImpl implements ProductService {
             throw new AccessDeniedException("User is not authenticated");
         }
 
-        if (product.isActive() == active) {
-            return;
-        }
 
-        product.setActive(active);
+        product.setStatus(Status.PENDING_REVIEW);
 
         productRepository.save(product);
     }
@@ -274,8 +268,9 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public void deleteProduct(Long id) {
 
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product not found with id: " + id, HttpStatus.NOT_FOUND));
+        Product product = productRepository
+                .findByIdAndStatus(id, List.of(Status.ACTIVE))
+                .orElseThrow(() -> new EntityNotFoundException("Product not found", HttpStatus.NOT_FOUND));
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (!(authentication.getPrincipal() instanceof User user)) {
             throw new AccessDeniedException("User is not authenticated");
@@ -283,37 +278,18 @@ public class ProductServiceImpl implements ProductService {
         if (!user.getId().equals(product.getCreatedBy().getId())) {
             throw new AccessDeniedException("You are not the owner of this product");
         }
-//        for (Attachment attachment : product.getAttachments()) {
-//
-//            attachmentService.deleteById(attachment.getId());
-//
-//        }
+
+        if (product.getStatus().equals(Status.DRAFT)) {
+            throw new IllegalActionException("This product is in draft state and cannot be deleted", HttpStatus.BAD_REQUEST);
+        }
 
         productRepository.delete(product);
 
     }
 
-    @Override
-    public void approveProduct(Long id) {
-
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product not found with id: " + id, HttpStatus.NOT_FOUND));
-
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (!(authentication.getPrincipal() instanceof User user)) {
-            throw new AccessDeniedException("User is not authenticated");
-        }
-        if (!user.getRole().equals(Role.ADMIN)) {
-            throw new AccessDeniedException("Only admins can approve products");
-        }
-
-        product.setIsApproved(true);
-        productRepository.save(product);
-
-    }
 
     @Override
-    public PageDTO<ProductDTO> getUserProducts(Integer page, Integer size) {
+    public PageDTO<ProductDTO> getUserProductsIsApprovedTrue(Integer page, Integer size) {
 
         Sort sort = Sort.by(LongIdAbstract.Fields.id);
         PageRequest pageRequest = PageRequest.of(page, size, sort);
@@ -323,7 +299,33 @@ public class ProductServiceImpl implements ProductService {
             throw new AccessDeniedException("User is not authenticated");
         }
 
-        Page<Product> byCreatedBy = productRepository.findByCreatedBy(user, pageRequest);
+        Page<Product> byCreatedBy = productRepository.findByCreatedByAndStatus(user, Status.ACTIVE, pageRequest);
+
+        return new PageDTO<>(
+                byCreatedBy.getContent().stream().map(productMapper::toDto).toList(),
+                byCreatedBy.getNumber(),
+                byCreatedBy.getSize(),
+                byCreatedBy.getTotalElements(),
+                byCreatedBy.getTotalPages(),
+                byCreatedBy.isLast(),
+                byCreatedBy.isFirst(),
+                byCreatedBy.getNumberOfElements(),
+                byCreatedBy.isEmpty()
+        );
+    }
+
+    @Override
+    public PageDTO<ProductDTO> getWaitingProducts(Integer page, Integer size) {
+
+        Sort sort = Sort.by(AbstractEntity.Fields.updatedAt);
+        PageRequest pageRequest = PageRequest.of(page, size, sort);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication.getPrincipal() instanceof User user)) {
+            throw new AccessDeniedException("User is not authenticated");
+        }
+
+        Page<Product> byCreatedBy = productRepository.findByCreatedByAndStatus(user, Status.PENDING_REVIEW, pageRequest);
 
         return new PageDTO<>(
                 byCreatedBy.getContent().stream().map(productMapper::toDto).toList(),
@@ -337,6 +339,84 @@ public class ProductServiceImpl implements ProductService {
                 byCreatedBy.isEmpty()
         );
 
+    }
+
+    @Override
+    public PageDTO<ProductDTO> getInactiveProducts(Integer page, Integer size) {
+
+        Sort sort = Sort.by(AbstractEntity.Fields.updatedAt);
+        PageRequest pageRequest = PageRequest.of(page, size, sort);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication.getPrincipal() instanceof User user)) {
+            throw new AccessDeniedException("User is not authenticated");
+        }
+
+        Page<Product> byCreatedBy = productRepository.findByCreatedByAndStatus(user, Status.INACTIVE, pageRequest);
+
+        return new PageDTO<>(
+                byCreatedBy.getContent().stream().map(productMapper::toDto).toList(),
+                byCreatedBy.getNumber(),
+                byCreatedBy.getSize(),
+                byCreatedBy.getTotalElements(),
+                byCreatedBy.getTotalPages(),
+                byCreatedBy.isLast(),
+                byCreatedBy.isFirst(),
+                byCreatedBy.getNumberOfElements(),
+                byCreatedBy.isEmpty()
+        );
 
     }
+
+    @Override
+    public PageDTO<ProductDTO> getRejectedProducts(Integer page, Integer size) {
+
+        Sort sort = Sort.by(AbstractEntity.Fields.updatedAt);
+        PageRequest pageRequest = PageRequest.of(page, size, sort);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication.getPrincipal() instanceof User user)) {
+            throw new AccessDeniedException("User is not authenticated");
+        }
+
+        Page<Product> rejectedProducts = productRepository.findByCreatedByAndStatus(
+                user,
+                Status.REJECTED,
+                pageRequest
+        );
+
+        return new PageDTO<>(
+                rejectedProducts.getContent().stream().map(productMapper::toDto).toList(),
+                rejectedProducts.getNumber(),
+                rejectedProducts.getSize(),
+                rejectedProducts.getTotalElements(),
+                rejectedProducts.getTotalPages(),
+                rejectedProducts.isLast(),
+                rejectedProducts.isFirst(),
+                rejectedProducts.getNumberOfElements(),
+                rejectedProducts.isEmpty()
+        );
+
+    }
+
+    public PageDTO<ProductDTO> searchProducts(ProductFilterDTO filterDTO, Pageable pageable) {
+
+        Specification<Product> specification = ProductSpecification.filterBy(filterDTO);
+
+        Page<Product> all = productRepository.findAll(specification, pageable);
+
+        return new PageDTO<>(all.getContent().stream().map(productMapper::toDto).toList(),
+                all.getNumber(),
+                all.getSize(),
+                all.getTotalElements(),
+                all.getTotalPages(),
+                all.isLast(),
+                all.isFirst(),
+                all.getNumberOfElements(),
+                all.isEmpty()
+        );
+
+    }
+
+
 }
